requirements.txt
opencv-python>=4.10.0
Pillow>=10.0.0
numpy>=1.26.0
PyQt6>=6.7.0
# Optional, for AI-based background removal if you need it later
rembg>=2.0.0


